[ALPS07806044] iommu/mediatek: optimize iommu debug for smmu
[Details]
The new chip needs to add the smmu debug mechanism, so some optimization
and adjustments need to be made to the existing iommu debug mechanism.

[solution]
1.Split larb port debug info definition into iommu_port.h
  to avoid iommu_debug.c single file too big;
2.tab_id is smmu hardware id For smmu, so adjust tab_id
  to u64 for adapt smmu;
3.Remove unnecessary old chip larb port debug info;
4.Remove unnecessary iommu_iova_dbg.h file;
5.Other small debug mechanism optimization and adjustment.

MTK-Commit-Id: 4a27dc70b82fcb6a8402346fcbc3a7a00cc36de1

CR-Id: ALPS07806044
Feature: iommu_kernel

+++ drivers/misc/mediatek/iommu/iommu_debug.h
@@ -61,9 +61,9 @@ int mtk_iommu_register_fault_callback(int port,
 
 /* port: comes from "include/dt-binding/memort/mtxxx-larb-port.h" */
 int mtk_iommu_unregister_fault_callback(int port, bool is_vpu);
-void mtk_iova_map(int tab_id, u64 iova, size_t size);
-void mtk_iova_unmap(int tab_id, u64 iova, size_t size);
-void mtk_iova_map_dump(u64 iova, u32 tab_id);
+void mtk_iova_map(u64 tab_id, u64 iova, size_t size);
+void mtk_iova_unmap(u64 tab_id, u64 iova, size_t size);
+void mtk_iova_map_dump(u64 iova, u64 tab_id);
 void mtk_iommu_tlb_sync_trace(u64 iova, size_t size, int iommu_ids);
 void mtk_iommu_pm_trace(int event, int iommu_id, int pd_sta,
 	unsigned long flags, struct device *dev);
